backend: look for OVF store disks only in data domains
OVF store disks exist only in data domains.
Thus, when we attach a storage domain to a dc, we should look for them
only if it's a data domain, and with that prevent irrelevant warning
messages from being logged.

Bug-Url: https://bugzilla.redhat.com/1305343

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/storage/pool/AddStoragePoolWithStoragesCommand.java
@@ -245,23 +245,23 @@ public class AddStoragePoolWithStoragesCommand<T extends StoragePoolWithStorages
 
     private void registerOvfStoreDisks() {
         for (final Guid storageDomainId : getParameters().getStorages()) {
-            resetOvfStoreDisks();
-            final List<OvfEntityData> unregisteredEntitiesFromOvfDisk =
-                    getEntitiesFromStorageOvfDisk(storageDomainId, getStoragePool().getId());
-            TransactionSupport.executeInNewTransaction(() -> {
-                List<DiskImage> ovfStoreDiskImages = getAllOVFDisks(storageDomainId, getStoragePool().getId());
-                registerAllOvfDisks(ovfStoreDiskImages, storageDomainId);
+            if (getStorageDomainStaticDao().get(storageDomainId).getStorageDomainType().isDataDomain()) {
+                resetOvfStoreDisks();
+                TransactionSupport.executeInNewTransaction(() -> {
+                    List<DiskImage> ovfStoreDiskImages = getAllOVFDisks(storageDomainId, getStoragePool().getId());
+                    registerAllOvfDisks(ovfStoreDiskImages, storageDomainId);
 
-                // Update unregistered entities
-                for (OvfEntityData ovf : unregisteredEntitiesFromOvfDisk) {
-                    getUnregisteredOVFDataDao().removeEntity(ovf.getEntityId(), storageDomainId);
-                    getUnregisteredOVFDataDao().saveOVFData(ovf);
-                    log.info("Adding OVF data of entity id '{}' and entity name '{}'",
-                            ovf.getEntityId(),
-                            ovf.getEntityName());
-                }
-                return null;
-            });
+                    // Update unregistered entities
+                    for (OvfEntityData ovf : getEntitiesFromStorageOvfDisk(storageDomainId, getStoragePool().getId())) {
+                        getUnregisteredOVFDataDao().removeEntity(ovf.getEntityId(), storageDomainId);
+                        getUnregisteredOVFDataDao().saveOVFData(ovf);
+                        log.info("Adding OVF data of entity id '{}' and entity name '{}'",
+                                ovf.getEntityId(),
+                                ovf.getEntityName());
+                    }
+                    return null;
+                });
+            }
         }
     }
